Refactor toast import paths and remove obsolete use-toast file for improved code organization
- Updated import paths for the toast utility across multiple components to use a consistent naming convention, enhancing clarity and maintainability.
- Deleted the obsolete `use-toast.ts` file, streamlining the codebase and reducing complexity.
